Node, Slot, Dice communication work
The communication between Node, slot, and dice was stablish and worked on.
Most of the logic is set.

To finish the interaction between the 3 of them, other scripts are missing.

the foundation for resource addition is set, and events can now be imported to affect slots.

Communicating results from the dice up the chain is also done.

Camera movement, and dragging objects was worked on but remove due to problems, and will be re introduced tomorrow